// samples/document-approval/web/src/app/generated/rulegate.ts
// <auto-generated />
// Generated by RuleGate. Do not edit.

export const RuleGateIdentifiers = {
  policies: {
    documentApprove: 'document-approve',
    documentCreate: 'document-create',
    documentRead: 'document-read',
    documentReject: 'document-reject',
    documentSubmit: 'document-submit',
    documentUpdate: 'document-update',
    documentsList: 'documents-list',
  },
  permissions: {
    docCreate: 'DOC.CREATE',
    docRead: 'DOC.READ',
    docUpdate: 'DOC.UPDATE',
    wflApprove: 'WFL.APPROVE',
    wflReject: 'WFL.REJECT',
    wflStart: 'WFL.START',
  },
  roles: {
    keycloakRealmApprover: 'keycloak:realm:APPROVER',
  },
  resourceTypes: {
    document: 'document',
  },
  actions: {
    approve: 'approve',
    create: 'create',
    list: 'list',
    read: 'read',
    reject: 'reject',
    submit: 'submit',
    update: 'update',
  },
} as const;

export type RuleGatePolicy =
  (typeof RuleGateIdentifiers.policies)[keyof typeof RuleGateIdentifiers.policies];
export type RuleGatePermission =
  (typeof RuleGateIdentifiers.permissions)[keyof typeof RuleGateIdentifiers.permissions];
export type RuleGateRole =
  (typeof RuleGateIdentifiers.roles)[keyof typeof RuleGateIdentifiers.roles];
export type RuleGateResourceType =
  (typeof RuleGateIdentifiers.resourceTypes)[keyof typeof RuleGateIdentifiers.resourceTypes];
export type RuleGateAction =
  (typeof RuleGateIdentifiers.actions)[keyof typeof RuleGateIdentifiers.actions];